Move lambda generation to a new class
More refactoring by moving the lambda byte code generation code to a separate class.
